Use validator set and consensus params given by initialize_chain
Since the genesis validator set and the genesis consensus params are
obtained from initialize_chain, they should be removed from Scheme.
initialize_chain is called when initializing a client. State is updated
after then.
run_node procedure is reordered, since we have to use network_id to
initialize email alarm.
